\name{pbc}
\alias{pbc}
\docType{package}

\title{Package of Percentile-based Control Charts}

\description{Create an object of class \code{'pbcc'} to perform statistical quality control based on percentile-based designs (PL) approach. This object may then be used to plot percentile-based Shewhart charts either individual or joint and summary statistics with guaranted in-control performace. Also, it determine the optimal parmeters n (sample size), h (sampling inetrval), k (control chart constant) for average chart, lp and up (lower and upper percentiles of the relative range/std.dev/variance distribution) for Statistcal Design of Percentile-based control charts.
}

\seealso{\code{\link{pbcc}}, \code{\link{odpbc}}, \code{\link{summary.pbcc}}, \code{\link{plot.pbcc}}.}

\author{Aamir Saghir, Zsolt T. Kosztyan*

e-mail: kzst@gtk.uni-pannon.hu}

\references{
Faraz A, Saniga E, Montgomery D. (2019). Percentile-based control charts design with an application to Shewhart Xbar and S2 control charts. Quality and Reliability Engineering International, 35(1); 116-126.
}

\keyword{package}
