\name{pbcc}
\alias{pbcc}
\docType{package}

\title{Package of Percentile-based Control Charts}

\description{Calculate the optimal parmeters n (sample size), h (sampling inetrval) and l (percentile of chi-square distribution) for Statistcal Design of Percentile-based control charts.
}

\seealso{\code{\link{pbs2}}, \code{\link{pbxbar}}, \code{\link{pbxbars2}}, \code{\link{summary.pbcc}}.}

\author{Aamir Saghir, Zsolt T. Kosztyan*

e-mail: kzst@gtk.uni-pannon.hu}

\references{
Faraz A, Saniga E, Montgomery D. (2019). Percentile-based control charts design with an application to Shewhart Xbar and S2 control charts. Quality and Reliability Engineering International, 35(1); 116-126.
}

\keyword{package}
